Compiler Warning [-Wmismatched-new-delete]
This warning are false positive, free() is a C library function that can also be used in C++.
Replace it with "delete" which is a C++ keyword and the compiler will stop complaining.
free() frees memory but doesn't call Destructor of a class
whereas "delete" frees the memory and also calls the Destructor of the class.
And "delete" is faster than free() because an operator is always faster than a function.